backend/requirements.txt
# Web backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML and Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.26.4
Pillow==10.4.0
tensorflow==2.15.0

# Audio processing
pyaudio==0.2.11
speechrecognition==3.10.0
whisper==1.1.10

# Utilities
pydantic==2.9.2
httpx==0.25.2
celery==5.3.4
redis==5.0.1

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
